handle merge conflict
